fix(optimum): allow non-fp32 dtypes for models that support them (#800)

+++ vllm_rbln/model_executor/models/optimum/optimum_attention/optimum_attention_strategy.py
@@ -259,7 +259,9 @@ class HybridAttentionImageStrategy(
         assert input_ids is not None
 
         if is_prompt:
-            attention_mask = (input_ids != self.pad_token_id).to(torch.int64).squeeze(0)
+            # Not an input of the compiled graph: optimum-rbln only uses this
+            # mask to drop padding from the prefill inputs.
+            attention_mask = (input_ids != self.pad_token_id).squeeze(0)
         else:
             get_extra_values_fn = lambda entry: (
                 entry.pad_len,
